% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTD.R
\name{calc_annual_bi_ratio}
\alias{calc_annual_bi_ratio}
\title{Calculate annual BI ratio}
\usage{
calc_annual_bi_ratio(s, i, conversion = c("Sum", "Average"))
}
\arguments{
\item{s}{ts or mts object annual benchmark(s)}

\item{i}{ts or mts object infra-annual indicator(s)}

\item{conversion}{a character object indicating the type of conversion}
}
\value{
ts or mts object annual BI ratio(s)
}
\description{
Calculate annual BI ratio
}
