% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_cv_tramo}
\alias{calc_cv_tramo}
\title{Time series cross validation for TRAMO}
\usage{
calc_cv_tramo(s, h, min.origin = 10)
}
\arguments{
\item{s}{\code{"\link{ts}"} object}

\item{h}{forecast horizon}

\item{min.origin}{minimum number of periods considered to start cross-validation process. Note that 'tramo.forecast' only works with nobs >= 10}
}
\value{
cross-validation errors
}
\description{
The function computes the cross validation errors for each period when using TRAMO as forecasting method
}
