% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{CalcBIratio}
\alias{CalcBIratio}
\title{Calculate annual BI ratio}
\usage{
CalcBIratio(s, i, conversion = c("Sum", "Average"))
}
\arguments{
\item{s}{ts or mts object annual benchmark(s)}

\item{i}{ts or mts object infra-annual indicator(s)}

\item{conversion}{a character object indicating the type of conversion}
}
\value{
ts or mts object annual BI ratio(s)
}
\description{
Calculate annual BI ratio
}
\examples{

data(nbb_data)
B1G_Y_data <- nbb_data$B1G_Y_data
TURN_Q_data <- nbb_data$TURN_Q_data

# convert data to ts object
B1G_Y_data_ts <- ts(B1G_Y_data[,-1], frequency = 1, start = c(2009,1))
TURN_Q_data_ts <- ts(TURN_Q_data[,-1], frequency = 4, start = c(2009,1))

CalcBIratio(B1G_Y_data_ts, TURN_Q_data_ts) #colnames of \code{s} is used for the output

}
