% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{CalcCVRMSE}
\alias{CalcCVRMSE}
\title{Calculate RMSE of cross-validation errors for a selected method}
\usage{
CalcCVRMSE(
  s,
  model = c("rw", "mean", "gmeang", "tramo", "holtd", "regl", "arima"),
  fh = 1,
  wd = 5,
  min.origin = 10
)
}
\arguments{
\item{s}{\code{"\link{ts}"} object}

\item{model}{forecasting model. 'rw' = random walk (naive) method, 'mean' = average method over a rolling window, ...}

\item{fh}{forecast horizon}

\item{wd}{size of the rolling window for the functions 'mean' and 'gmeang'}

\item{min.origin}{minimum number of periods considered to start cross-validation process. Note that TRAMO needs a minimum of 10 periods.}
}
\value{
a vector with RMSE of cross-validation errors for the choosen model
}
\description{
Calculate RMSE of cross-validation errors for a selected method
}
\examples{
data(nbb_data)
B1G_Y_data_ts <- ts(nbb_data$B1G_Y_data[,-1], frequency = 1, start = c(2009,1))

#select one series
s_CE <- B1G_Y_data_ts[,1]

CalcCVRMSE(s_CE, model = "holtd")

}
