% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{CalcCVRMSEOfSelectedModels}
\alias{CalcCVRMSEOfSelectedModels}
\title{Calculate RMSE of cross validation errors for several forecasting models}
\usage{
CalcCVRMSEOfSelectedModels(s)
}
\arguments{
\item{s}{\code{"\link{ts}"} object annual time series}
}
\value{
matrix
}
\description{
Calculate RMSE of cross validation errors for several forecasting models
}
\examples{
data(nbb_data)
B1G_Y_data_ts <- ts(nbb_data$B1G_Y_data[,-1], frequency = 1, start = c(2009,1))

#select one series
s_CE <- B1G_Y_data_ts[,1]

CalcCVRMSEOfSelectedModels(s_CE)

}
