% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedDenton.R
\name{forecast_series}
\alias{forecast_series}
\title{Forecast annual time series according to a selected method}
\usage{
forecast_series(
  s,
  model = c("rw", "mean", "gmeang", "tramo", "holtd", "arima", "regl"),
  fh = 2,
  wd = 5
)
}
\arguments{
\item{s}{\code{"\link{ts}"} object}

\item{model}{model used to forecast. 'rw' = random walk (naive) method, 'mean' = average method over a rolling window, ...}

\item{fh}{forecast horizon}

\item{wd}{size of the rolling window for the functions 'mean' and 'gmeang' (ignored when other method)}
}
\value{
vector with forecasts
}
\description{
Forecast annual time series according to a selected method
}
