% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{ForecastAnnualBIRatios}
\alias{ForecastAnnualBIRatios}
\title{Automatic forecast of annual BI ratio's}
\usage{
ForecastAnnualBIRatios(bi_ts, critical.value, bi_f_manual = NULL)
}
\arguments{
\item{bi_ts}{mts object annual BI ratio's to forecast}

\item{critical.value}{a numeric object threshold value to use in the
automatic selection of  the model}

\item{bi_f_manual}{a data.frame object with 3 columns: 'series_name',
'f_T1' and 'f_T2' containing manual forecast of annual
 BI ratio for years T+1 and T+2. If the data.frame is
 empty or if f_bi_manual = NULL, the results of the
 automatic selection of forecasting model is used by
 default for each series.}
}
\value{
list of 2 elements with primary and alternative forecasts
}
\description{
Automatic forecast of annual BI ratio's
}
\examples{
data(nbb_data)
B1G_Y_data <- nbb_data$B1G_Y_data
TURN_Q_data <- nbb_data$TURN_Q_data

# convert data to ts object
B1G_Y_data_ts <- ts(B1G_Y_data[,-1], frequency = 1, start = c(2009,1))
TURN_Q_data_ts <- ts(TURN_Q_data[,-1], frequency = 4, start = c(2009,1))

# compute annual Bi ration
bi_ts <- CalcBIratio(B1G_Y_data_ts, TURN_Q_data_ts)

# forecast of annual BI ratio (only used when enhanced=TRUE)
data(table_rw)
f_bi_ts <- ForecastAnnualBIRatios(bi_ts, critical.value = table_rw[series_length == nrow(bi_ts), q.95])

}
