% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{ForecastSeries}
\alias{ForecastSeries}
\title{Forecast annual time series according to a selected method}
\usage{
ForecastSeries(
  s,
  model = c("rw", "mean", "gmeang", "tramo", "holtd", "arima", "regl"),
  fh = 2,
  wd = 5
)
}
\arguments{
\item{s}{\code{"\link{ts}"} object}

\item{model}{model used to forecast. 'rw' = random walk (naive) method, 'mean' = average method over a rolling window, ...}

\item{fh}{forecast horizon}

\item{wd}{size of the rolling window for the functions 'mean' and 'gmeang' (ignored when other method)}
}
\value{
vector with forecasts
}
\description{
Forecast annual time series according to a selected method
}
\examples{
data(nbb_data)
B1G_Y_data_ts <- ts(nbb_data$B1G_Y_data[,-1], frequency = 1, start = c(2009,1))

#select one series
s_CE <- B1G_Y_data_ts[,1]

ForecastSeries(s_CE, model = "holtd")

}
