% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbDenton.R
\name{mbdenton}
\alias{mbdenton}
\title{Model-based Denton}
\usage{
mbdenton(
  indicator,
  benchmark,
  outliers = NULL,
  outliers.intensity = 10,
  manual_disagBI = NULL,
  conversion = c("Sum", "Average"),
  series_name = NULL,
  include_warnings = TRUE
)
}
\arguments{
\item{indicator}{monthly or quarterly indicator. Must be a ts object.}

\item{benchmark}{annual benchmark. Must be a ts object.}

\item{outliers}{level shift(s) in the Benchmark-to-Indicator ratio. Must be
specified as a numeric vector with the outlier periods in decimal dates.
Default is NULL (no outlier).}

\item{outliers.intensity}{intensity of the  outlier, defined as the relative
value of the 'innovation variances' (1= normal situation). Must be
specified as a numeric vector corresponding to the outlier periods
specified in the outliers argument. Default is 10 for each outlier.}

\item{manual_disagBI}{Disaggregated BI ratio that the user wants to fix. Must
be as a matrix of two columns called 'period' and 'bi_ratio'. The periods
must be in decimal. Default is NULL (no intervention). If not NULL, they
must be provided for each quarter/month of the year.}

\item{conversion}{type of consistency between the annual benchmarks and the
infra-annual indicators.}

\item{series_name}{Character. Name of the series. Optional argument used to
customize warning message if any.}
}
\value{
a list
}
\description{
Proportional First Difference Denton method put in State Space form and
including the possibility to add outliers (shifts in the disaggregated BI
ratio) and manual figures for the disaggregated BI ratio.
}
\examples{
Y<-rjd3toolkit::aggregate(rjd3toolkit::Retail$RetailSalesTotal, 1)
x<-rjd3toolkit::aggregate(rjd3toolkit::Retail$FoodAndBeverageStores, 4)

outliers<-c(2008.75,2009.00)
outliers.intensity<-c(10,5)
manual_disagBI<-matrix(cbind(c(2008.00,2008.25,2008.50,2008.75),
                             c(7.246695,7.246695,7.246695,6.0)),
                       ncol=2,
                       dimnames=list(NULL,c("period","bi_ratio")))
rslt<-mbdenton(x,Y,outliers,outliers.intensity,manual_disagBI,conversion="Sum")

}
