% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{multiTD}
\alias{multiTD}
\title{Multiprocessing NBB-customized temporal disaggregation of time series}
\usage{
multiTD(
  benchmarks,
  indicators,
  config,
  conversion = c("Sum", "Average"),
  f.score.quantile = c("q.95", "q.90", "q.80", "q.99", "q.999"),
  simpl.e.T1 = FALSE,
  path.xlsx = ""
)
}
\arguments{
\item{benchmarks}{set of benchmarks time series that will be disaggregated.
It can be an object of class mts or data.frame.
In the second case, the first column should be in a format
that can be translated into a Date object.
Only annual benchmarks are handled.}

\item{indicators}{set of corresponding indicators time series used for
temporal disaggregation. As for the benchmarks, it can be
an object of class mts or data.frame. In the second
case, the first column should be in a format that can be
translated into a Date object. If no indicator exists, the
indicator series should be defined as NA or any constant
so that the number of columns is identical to this of the
benchmarks (one indicator by benchmark). Name and order of
the columns must be identical to those of the benchmarks.
Only quarterly or monthly indicators are handled.}

\item{config}{an object of the class 'nbb.dsc.td.multiproc.config'.
The object contains a list with the required information to
perform temporal disaggregation of each series.}

\item{conversion}{type of consistency between the annual benchmarks and the
infra-annual indicators}

\item{f.score.quantile}{quantile to consider when choosing between a random
walk process and another model to forecast the annual
BI ratio. A lower value of the quantile will increase
the propensity to select another model and vice-versa.}

\item{simpl.e.T1}{a boolean object. By default, when simpl.e.T1 = FALSE, the
benchmarks series are automatically extended whenever the
indicators series cover the year T+1. If simpl.e.T1 = TRUE,
the simple extrapolation is still used. That allows to a
void revision of the past of the series. Note that this is
only relevant when the range of the indicators series
covers exactly the year T+1. If it goes beyond
(i.e. covering a part of year T+2), the argument is ignored
and the benchmarks series are extended.}

\item{path.xlsx}{a character object containing the path of the XLSX file to
generate with the main results. If the path is en empty
string, no XLSX file is created.}
}
\value{
an object of class "nbb.dsc.td.multiproc.output"
}
\description{
Perform multi-processing temporal disaggregation of an annual to quarterly
or monthly set of time series. The focus is on the (enhanced) Denton PFD
method defined in State Space form. Hence, the user can also add outliers
(level shift) in the Benchmark-to-Indicator (BI) ratio and import
infra-annual BI ratio manually for some periods. When enhanced Denton is
used, an automatic procedure exists to select the method for the forecast
of the annual BI ratio but the user can also change the forecasts manually
if necessary.In addition to Denton PDF method, Chow-Lin method and its
variants can also be used. Either the sum or the average consistency between
the annual benchmarks and the quarterly or monthly indicators are handled.
}
\examples{
data(nbb_data)
B1G_Y_data <- nbb_data$B1G_Y_data
TURN_Q_data <- nbb_data$TURN_Q_data
colnames(B1G_Y_data) <- colnames(TURN_Q_data) <- c("DATE", "CE", "FF", "HH") # mandatory to match colnames

data(table_rw) # table of critical values for the automatic selection of the forecasting model of annual BI ratio with 'eDenton'

# define config
myconfig <- setConfig_default() # you can change any default argument you want. See ?setConfig_default
# or alternatively:
\dontrun{
myconfig <- set_nbbTDConfig_FromXLSX(file = "data/xxx.xlsx")
}

res <- multiTD(B1G_Y_data, TURN_Q_data, myconfig)

}
