% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTD.R
\name{multiTD_fromXLSX}
\alias{multiTD_fromXLSX}
\title{Multiprocessing temporal disaggregation of time series with input coming
from a structured XLSX file}
\usage{
multiTD_fromXLSX(
  path_data,
  forecastBI = c("none", "auto", "userdefined+none", "userdefined+auto"),
  forecastBI.quantile = c("q.95", "q.80", "q.90", "q.99", "q.999"),
  freezeT1 = FALSE,
  conversion = c("Sum", "Average"),
  path_output = NULL
)
}
\arguments{
\item{path_data}{a string containing the path of the input XLSX file.}

\item{forecastBI}{Forecast of the annual BI ratio to consider for the model-
based Denton. If "none" is selected, no forecast is considered ('standard'
model-based Denton). If "auto" is selected, an automatic forecast procedure
of the annual BI ratio will be used. "userdefined+none" and
"userdefined+auto" allows the user to specify the forecast of the annual BI
ratio to consider for some series.}

\item{forecastBI.quantile}{quantile to consider when choosing between a
random walk process and another model to forecast the annual BI ratio. A
lower value of the quantile will increase the propensity to select another
model and vice-versa.}

\item{freezeT1}{a boolean object. By default, when freezeT1 = FALSE, the
benchmarks series are automatically extended whenever the indicators series
cover the year T+1. If freezeT1 = TRUE, the simple extrapolation is still
used. That allows to avoid revision of the past of the series. Note that
this is only relevant when the range of the indicators series covers
exactly the year T+1. If it goes beyond (i.e. covering a part of year T+2),
the argument is ignored and the benchmarks series are extended.}

\item{conversion}{type of consistency between the annual benchmarks and the
infra-annual indicators}

\item{path_output}{a string containing the path of the output (which is also
an XLSX file) containing the results. If the path is NULL, no file is
created.}
}
\value{
an object of class "nbb.multiTD.output"
}
\description{
See ?multiTD for more description. See vignette or the template for
explanation about the required structure of the XLSX file.
}
\examples{
\dontrun{
file.name<-"myinput.xlsx"
res<-multiTD_fromXLSX(path_data = "input_to_R.xlsx",
                      forecastBI = "userdefined+auto",
                      forecastBI.quantile = "q.95",
                      freezeT1 = FALSE,
                      conversion = "Sum",
                      path_output = "output_R.xlsx")
}

}
