% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{multiTDCLvar}
\alias{multiTDCLvar}
\title{Multiprocessing temporal disaggregation of time series with
Chow-Lin method and its variants}
\usage{
multiTDCLvar(
  benchmarks_ts,
  indicators_ts,
  models,
  conversion = c("Sum", "Average")
)
}
\arguments{
\item{benchmarks_ts}{set of benchmarks time series that will be disaggregated.
It must be a mts object.}

\item{indicators_ts}{set of corresponding indicators time series used for
temporal disaggregation. It must be a mts object.
If no indicator exists, the indicator series should
be defined as NA or any constant so that the number
of columns is identical to this of the benchmarks.
Name and order of the columns must be identical to those
of the benchmarks too.}

\item{models}{a character vector with the name of the model to use for each
series in the order of the columns in 'benchmarks_ts'. There
are 3 possibilities of models: 'chow-lin', 'fernandez'
or 'litterman'.}

\item{conversion}{type of consistency between the annual benchmarks and the
infra-annual indicators}
}
\value{
an object of class 'nbb.dsc.td.multiproc.clvar.output'
}
\description{
Perform temporal disaggregation of an annual to quarterly or monthly set of
time series using Chow-Lin method or its variants.
}
\details{
Either the sum or the average consistency between the annual benchmarks and
the quarterly or monthly indicators are handled.
}
\examples{
data(nbb_data)
B1G_Y_data <- nbb_data$B1G_Y_data
TURN_Q_data <- nbb_data$TURN_Q_data
colnames(B1G_Y_data) <- colnames(TURN_Q_data) <- c("DATE", "CE", "FF", "HH") # mandatory to match colnames

# convert data to ts object
B1G_Y_data_ts <- ts(B1G_Y_data[,-1], frequency = 1, start = c(2009,1))
TURN_Q_data_ts <- ts(TURN_Q_data[,-1], frequency = 4, start = c(2009,1))

# define models
models <- c("chow-lin", "fernandez", "chow-lin")

res <- multiTDCLvar(B1G_Y_data_ts, TURN_Q_data_ts, models, conversion = "Sum")

}
