% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbb_data}
\alias{nbb_data}
\title{nbb_data}
\format{list of two data.frames containing the example data}
\source{
real data examples taken from Belgian National Accounts.
}
\usage{
nbb_data
}
\description{
Real data examples of annual benchmark series and corresponding quarterly
indicators taken from Belgian National Accounts. The letters used for the
column names refers to the official clissification NACE A38.
}
\keyword{datasets}
