% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbTD.R
\docType{package}
\name{nbbTD}
\alias{nbbTD}
\alias{nbbTD-package}
\title{Multiprocessing NBB-customized temporal disaggregation of time series}
\description{
Perform multi-processing temporal disaggregation of an annual to quarterly
or monthly set of time series. The focus is on the (enhanced) Denton PFD
method defined in State Space form. Hence, the user can also add outliers
(level shift) outliers (level shift) in the Benchmark-to-Indicator (BI) ratio
and import infra-annual BI ratio manually for some periods. In addition to
Denton PDF method, Chow-Lin method and its variants can also be used.
}
\author{
Corentin Lemasson \email{corentin.lemasson@nbb.be}
}
