% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.nbb.dsc.td.multiproc.output.denton}
\alias{plot.nbb.dsc.td.multiproc.output.denton}
\title{Plot function for objects of sub-classes 'nbb.dsc.td.multiproc.output.denton'}
\usage{
\method{plot}{nbb.dsc.td.multiproc.output.denton}(x, series_name = "", ...)
}
\arguments{
\item{x}{an object of class 'nbb.dsc.td.multiproc.output.denton'}

\item{series_name}{a character object with the name of the series}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Plot function for objects of sub-classes 'nbb.dsc.td.multiproc.output.denton'
}
