% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{plot.nbb.multiTD.output}
\alias{plot.nbb.multiTD.output}
\title{Plot function for objects of class 'nbb.multiTD.output'}
\usage{
\method{plot}{nbb.multiTD.output}(x, series_name = NULL, ...)
}
\arguments{
\item{x}{an object of class 'nbb.multiTD.output'}

\item{series_name}{a character object with the name of the series. Must be
kept to NULL for single processing.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Plot function for objects of class 'nbb.multiTD.output'
}
