% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_annual_bi}
\alias{plot_annual_bi}
\title{Plot annual BI ratio (Shiny)}
\usage{
plot_annual_bi(bi_ts, f_bi_ts, series_name, scaled = FALSE, ...)
}
\arguments{
\item{bi_ts}{set of Bi ratio's. It must be a mts object.}

\item{f_bi_ts}{set of forecasted Bi ratio's. It must be a mts object.}

\item{series_name}{a character object. Name of the series to plot}

\item{scaled}{a boolean object. If scaled = TRUE, the BI ratio is scaled on
to facilitate the interpretation of the change from year to year}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Plot annual BI ratio (Shiny)
}
