% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_bi_compare_manual}
\alias{plot_bi_compare_manual}
\title{Plot (e)Denton: comparison BI ratio with/without manual input (Shiny)}
\usage{
plot_bi_compare_manual(res_denton, series_name)
}
\description{
Plot (e)Denton: comparison BI ratio with/without manual input (Shiny)
}
