% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_decomposition_td_series}
\alias{plot_decomposition_td_series}
\title{Plot decomposition of TD series (Shiny)}
\usage{
plot_decomposition_td_series(td_series, decompositions, series_name)
}
\description{
Plot decomposition of TD series (Shiny)
}
