% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_td_series}
\alias{plot_td_series}
\title{Disaggregated series with CI (Shiny)}
\usage{
plot_td_series(td_series, td_series_stderr, series_name)
}
\description{
Disaggregated series with CI (Shiny)
}
