% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_td_series_ci}
\alias{plot_td_series_ci}
\title{Plot (e)Denton TD series with confidence intervals (Shiny)}
\usage{
plot_td_series_ci(x, series_name)
}
\arguments{
\item{x}{an object of class 'nbb.dsc.td.multiproc.output.denton'}

\item{series_name}{a character object with the name of the series}
}
\description{
Plot (e)Denton TD series with confidence intervals (Shiny)
}
