% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{plot_td_series_decomp}
\alias{plot_td_series_decomp}
\title{Plot decomposition of CLvar TD series (Shiny)}
\usage{
plot_td_series_decomp(x, series_name)
}
\arguments{
\item{x}{an object of class 'nbb.dsc.td.multiproc.output.clvar'}

\item{series_name}{a character object with the name of the series}
}
\description{
Plot decomposition of CLvar TD series (Shiny)
}
