% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{PlotAnnualBIratio}
\alias{PlotAnnualBIratio}
\title{Plot annual BI ratio for a specific series}
\usage{
PlotAnnualBIratio(bi_ts, f_bi_ts, series_name, scaled = FALSE, ...)
}
\arguments{
\item{bi_ts}{set of Bi ratio's. It must be a mts object.}

\item{f_bi_ts}{set of forecasted Bi ratio's. It must be a mts object.}

\item{series_name}{a character object. Name of the series to plot}

\item{scaled}{a boolean object. If scaled = TRUE, the BI ratio is scaled on
to facilitate the interpretation of the change from year to year}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Plot annual BI ratio for a specific series
}
\examples{
data(nbb_data)
B1G_Y_data <- nbb_data$B1G_Y_data
TURN_Q_data <- nbb_data$TURN_Q_data
colnames(B1G_Y_data) <- colnames(TURN_Q_data) <- c("DATE", "CE", "FF", "HH") # mandatory to match colnames

# convert data to ts object
B1G_Y_data_ts <- ts(B1G_Y_data[,-1], frequency = 1, start = c(2009,1))
TURN_Q_data_ts <- ts(TURN_Q_data[,-1], frequency = 4, start = c(2009,1))

# forecast of annual BI ratio (only used when enhanced=TRUE)
bi_ts <- CalcBIratio(B1G_Y_data_ts, TURN_Q_data_ts)
data(table_rw)
f_bi_ts <- ForecastAnnualBIRatios(bi_ts, critical.value = table_rw[series_length == nrow(bi_ts), q.95])$f_bi_ts

PlotAnnualBIratio(bi_ts, f_bi_ts, series_name = "HH", scaled = TRUE)

}
