% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{runShiny}
\alias{runShiny}
\title{Run Shiny App on the results}
\usage{
runShiny(res)
}
\arguments{
\item{res}{an object of class 'nbb.dsc.td.multiproc.output'}
}
\description{
Visualization of the results of class 'nbb.dsc.td.multiproc.output'
}
