% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setConfig_default}
\alias{setConfig_default}
\title{Creating a config object to be used as input in the functions 'multiTD',
or as part in the functions 'multiTDDenton', 'multiTDCLvar'.}
\usage{
setConfig_default(
  model = NULL,
  f_bi_manual = NULL,
  bi_q_outliers = NULL,
  bi_q_manual = NULL
)
}
\arguments{
\item{model}{a data.frame object with 2 columns: 'series_name' and
'model' containing the TD model to use for each series.
Currently, the following models are handled: 'edenton',
'denton', 'chow-lin', 'fernandez', 'litterman'.
The data.frame should contain at least one row. If you attempt
to use the same model for all series, you can use the keyword
'_ALL_'. If you do not define a model for each series,
'edenton' is used by default for the unspecified series
with a warning. If model = NULL, a data.frame is created
with 'eDenton' defined for each series by default.}

\item{f_bi_manual}{a data.frame object with 3 columns: 'series_name',
'f_T1' and 'f_T2' containing manual forecast of annual
 BI ratio for years T+1 and T+2. If the data.frame is
 empty or if f_bi_manual = NULL, the results of the
 automatic selection of forecasting model is used by
 default for each series.
 Only relevant for 'eDenton' model. f_bi_manual is ignored
 when other models are used with a warning message.}

\item{bi_q_outliers}{a data.frame object with 4 columns: 'series_name',
'year', quarter_month' and 'intensity' containing manual
outliers (and their intensity). If the data.frame is
empty or if bi_q_outliers = NULL, no outlier are considered.
The column 'quarter_month' must be filled with values
from 1 to 4 for quarterly data and with values from 1 to 12
for monthly data. The column 'intensity' allows the user
to set up the intensity of the outlier. The default value
is 10 and a value of 1 means no outlier. The scale
between 1 to 10 is not linear but exponential.}

\item{bi_q_manual}{a data.frame object with 14 columns: 'series_name',
'year', 'q1_m1', ..., 'q4_m4', 'm5', ..., 'm12' containing
manual BI ratios (on an infra-annual basis). If the
data.frame is empty or if bi_q_manual = NULL, no manual
BI ratios are considered. The columns 'm5' to 'm12' must
remain empty in case of quarterly data.}
}
\value{
returns an object of class 'nbb.dsc.td.multiproc.config'
}
\description{
Creating a config object to be used as input in the functions 'multiTD',
or as part in the functions 'multiTDDenton', 'multiTDCLvar'.
}
\examples{
model <- data.frame (series_name  = c("a","b","c"),
                     model = c("edenton","chow-lin","denton"))
f_bi_manual <- data.frame (series_name  = c("a","b","c"),
                           f_T1 = c(0.8,0.5,10),
                           f_T2 = c(0.82,0.5,9.9))
bi_q_outliers <- data.frame (series_name  = c("a","b","c"),
                             year = c(2020,2020,2016),
                             quarter_month = c(2,2,4),
                             intensity = c(10,10,10)) # keep intensity=10 by default
bi_q_manual <- data.frame(series_name  = c("a","b","c"),
                           year = c(2020,2020,2016),
                           q1_m1 = c(0.8,0.5,10),
                           q2_m2 = c(0.7,0.4,8),
                           q3_m3 = c(0.84,0.55,11),
                           q4_m4 = c(0.85,0.54,11), # all quarters/months of the year must be provided
                           m5 = c(NA,NA,NA), # fill with NA from here on for quarterly data
                           m6 = c(NA,NA,NA),
                           m7 = c(NA,NA,NA),
                           m8 = c(NA,NA,NA),
                           m9 = c(NA,NA,NA),
                           m10 = c(NA,NA,NA),
                           m11 = c(NA,NA,NA),
                           m12 = c(NA,NA,NA))

}
