% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setConfig_FromXLSX}
\alias{setConfig_FromXLSX}
\title{Creating a config object to be used as full input in the functions 'nbbTD',
or as part in the functions 'nbbTDDenton', 'nbbTDCLvar'.}
\usage{
setConfig_FromXLSX(file)
}
\arguments{
\item{file}{a character object with the path to the XLSX file containing
the config settings to consider. The XLSX file must respect a
specific structure. See help file of 'setConfig_default'
for more information.}
}
\value{
returns an object of class 'nbb.dsc.td.multiproc.config'
}
\description{
Creating a config object to be used as full input in the functions 'nbbTD',
or as part in the functions 'nbbTDDenton', 'nbbTDCLvar'.
}
\examples{
\dontrun{
myconfig <- setConfig_FromXLSX(file = "data/xxx.xlsx")
}

}
