% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.nbb.dsc.td.multiproc.output.denton}
\alias{summary.nbb.dsc.td.multiproc.output.denton}
\title{Summary function for objects of sub-classes 'nbb.dsc.td.multiproc.output.denton' and 'nbb.dsc.td.multiproc.output.clvar'}
\usage{
\method{summary}{nbb.dsc.td.multiproc.output.denton}(x, ...)
}
\arguments{
\item{x}{an object of class 'nbb.dsc.td.multiproc.output.denton' or 'nbb.dsc.td.multiproc.output.clvar'}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Summary function for objects of sub-classes 'nbb.dsc.td.multiproc.output.denton' and 'nbb.dsc.td.multiproc.output.clvar'
}
