% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_rw}
\alias{table_rw}
\title{table_rw data set}
\format{A data.table with 11 columns including series length and the
        critical value for various quantiles}
\source{
simulated data
}
\usage{
table_rw
}
\description{
Simulated data used to create table with critical value for the
automatic selection of the model for the forecasting of the annual BI ratio
}
\keyword{datasets}
