% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{ValidateAndFormatBIqManual}
\alias{ValidateAndFormatBIqManual}
\title{Validation and formatting of manual infra-annual BI ratios}
\usage{
ValidateAndFormatBIqManual(bi_q_manual, benchmarks_series_names, freq)
}
\arguments{
\item{bi_q_manual}{a data.frame or a data.table object with 14 columns: '
series_name', 'year', 'q1_m1', ..., 'q4_m4', 'm5', ...,
'm12' containing anual BI ratios on an infra-annual basis.}

\item{benchmarks_series_names}{a character vector containing the valid
the series names of the benchmarks series}

\item{freq}{a numeric object: 4 for quarterly data, 12 for monthly data}
}
\value{
data.table object
}
\description{
Validate and format initial data.table which include the manual infra-annual
BI ratios in a way that can be handled by the function 'multiTD', 'multiTDDenton'
and multiTDCLvar'. See help of the function 'set_multiTDConfig' for more info
about the structure of the input.
}
\examples{
# see ?setConfig_default for more details about the structure of the original input
bi_q_manual <- data.frame(series_name  = "CE",
                           year = 2020,
                           q1_m1 = 24.2,
                           q2_m2 = 23.5,
                           q3_m3 = 24.2,
                           q4_m4 = 24.2)
bi_q_manual_form <- ValidateAndFormatBIqManual(bi_q_manual, c("CE", "XXX", "YYY"), 4)

}
