% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_exported_functions.R
\name{ValidateAndFormatBIqOutliers}
\alias{ValidateAndFormatBIqOutliers}
\title{Validation and formatting of outliers in infra-annual BI ratios}
\usage{
ValidateAndFormatBIqOutliers(bi_q_outliers, benchmarks_series_names, freq)
}
\arguments{
\item{bi_q_outliers}{a data.frame or data.table object with 4 columns:
'series_name', 'year', quarter_month' and 'intensity'
containing manual outliers (and their intensity, set to
10 by default).}

\item{benchmarks_series_names}{a character vector containing the valid
the series names of the benchmarks series}

\item{freq}{a numeric object: 4 for quarterly data, 12 for monthly data}
}
\value{
data.table object
}
\description{
Validate and format initial data.table which include the infra-annual BI
outliers (or BI shifts) in a way that can be handled by the function 'multiTD',
'multiTDDenton' and multiTDCLvar'. See help of the function 'setConfig_default'
for more info about the structure of the input.
}
\examples{
# see ?setConfig_default for more details about the structure of the original input
bi_q_outliers <- data.frame(series_name  = "CE",
                            year = 2020,
                            quarter_month = 2,
                            intensity = 10) # keep intensity=10 by default
bi_q_outliers_form <- ValidateAndFormatBIqOutliers(bi_q_outliers, c("CE", "XXX", "YYY"), 4)

}
