% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTD.R
\name{multiTD}
\alias{multiTD}
\title{Multiprocessing temporal disaggregation of time series}
\usage{
multiTD(
  benchmarks,
  indicators,
  model = "mbDenton",
  outliers = NULL,
  disagBIfixed = NULL,
  forecastBI = c("none", "auto", "userdefined+none", "userdefined+auto"),
  forecastBI.values = NULL,
  forecastBI.quantile = c("q.95", "q.80", "q.90", "q.99", "q.999"),
  freezeT1 = FALSE,
  conversion = c("Sum", "Average"),
  path_xlsx = NULL
)
}
\arguments{
\item{benchmarks}{set of annual benchmark series to disaggregate. It can be
either a ts object or a data.frame. In the second case, the first column
should be in a format that can be translated into a Date object. Only
annual benchmarks are handled.}

\item{indicators}{set of monthly or quarterly indicators to use for the
disaggregation of the annual benchmark series. It can be either a ts object
or a data.frame. Indicators name must start with the name of the related
benchmark series. However, the name may be extended using an underscore,
for example when we'd like to use more than one indicator in a Chow-Lin
model. Only quarterly or monthly indicators are handled.}

\item{model}{Temporal disaggregation model to consider for each series.
Possible models are currently "mbdenton", "chow-lin", "fernandez" and
"litterman". The model to use can be a character string (when the same
model must be used for all series) or a list of structured definition of
the models. The name of the series must be submitted first. This must be
followed by an equal sign and the model's name to use for the series.}

\item{outliers}{Level shift(s) in the Benchmark-to-Indicator ratio for the
model-based Denton. Must be specified as a list of structured definition of
the outlier periods and their intensity for each series. The name of the
series must be submitted first. This must be followed by an equal sign and
the outliers description inside a vector (i.e., c(o1,o2,etc.)). The
outliers description must start with the period formatted as YYYYMMDD and
enclosed in quotation marks, then an equal sign and the intensity of the
outlier, defined as the relative value of the 'innovation variances' (1=
normal situation)}

\item{disagBIfixed}{Fix some disaggragated Benchmark-to-Indicator (BI) ratios
for the model-based Denton. They must be specified as a list of structured
definition of the periods of the disaggragated BI ratios and their values
for each series. The name of the series must be submitted first. This must
be followed by an equal sign and the disaggragated BI ratios inside a
vector (i.e., c(bi1,bi2,etc.)). The disaggragated BI ratio must start with
the period formatted as YYYYMMDD and enclosed in quotation marks, then an
equal sign and the value of the BI ratio at that period. When fixed
disaggregated BI ratio are provided, they must be provided for each
quarter/month of the year.}

\item{forecastBI}{Forecast of the annual BI ratio to consider for the model-
based Denton. If "none" is selected, no forecast is considered ('standard'
model-based Denton). If "auto" is selected, an automatic forecast procedure
of the annual BI ratio will be used. "userdefined+none" and
"userdefined+auto" allows the user to specify the forecast of the annual BI
ratio to consider for some series.}

\item{forecastBI.values}{Only used if userdefined forecast of the annual BI
ratio are brought in. Manual forecast must be specified as a list of
structured definition of the forecast values for each series. The name of
the series must be submitted first. This must be followed by an equal sign
and the forecasted values inside a vector of two values (i.e.,
c(f_y+1,f_y+2)). The forecasted values must start with "Y+1" and "Y+2"
respecively (enclosed in quotation marks), followed by an equal sign and
the value of the forecast.}

\item{forecastBI.quantile}{quantile to consider when choosing between a
random walk process and another model to forecast the annual BI ratio. A
lower value of the quantile will increase the propensity to select another
model and vice-versa.}

\item{freezeT1}{a boolean object. By default, when freezeT1 = FALSE, the
benchmarks series are automatically extended whenever the indicators series
cover the year T+1. If freezeT1 = TRUE, the simple extrapolation is still
used. That allows to avoid revision of the past of the series. Note that
this is only relevant when the range of the indicators series covers
exactly the year T+1. If it goes beyond (i.e. covering a part of year T+2),
the argument is ignored and the benchmarks series are extended.}

\item{conversion}{type of consistency between the annual benchmarks and the
infra-annual indicators}

\item{path_xlsx}{a character object containing the path of the XLSX file to
generate with the main results. If NULL (default), no XLSX file is created.}
}
\value{
an object of class "nbb.multiTD.output"
}
\description{
Perform multi-processing temporal disaggregation of an annual to quarterly or
monthly set of time series. Amongst the possible methods, the model-based
proportional first difference (PFD) Denton is defined in State Space form. It
offers the possibility to add outliers (level shift) in the
Benchmark-to-Indicator (BI) ratio and import infra-annual BI ratio manually
for some periods. The function also includes an automatic procedure to
forecast the annual BI ratio and takes this forecast into account in the
estimation of the out-of-sample periods. Instead of the automatic procedure,
forecasts of the annual BI ratio can also be defined by the user.
}
\examples{
# retail data, 2 benchmarks series to disaggregate
Y<-cbind(rjd3toolkit::aggregate(rjd3toolkit::retail$RetailSalesTotal, 1),
         rjd3toolkit::aggregate(rjd3toolkit::retail$ClothingStores, 1))
colnames(Y)<-c("retail","clothing")
x<-cbind(rjd3toolkit::retail$FoodAndBeverageStore,
         rjd3toolkit::retail$AutomobileDealers,
         rjd3toolkit::retail$WomensClothingStores)
colnames(x)<-c("retail_food", "retail_cars", "clothing_womens")

# Example 1: Chow-Lin (with 2 indicators) and standard Denton
rslt1<-multiTD(Y, x, model=list(retail="Chow-Lin", clothing="mbDenton"))

# Example 2a: Model-based Denton with outliers (i.e., LS in the BI ratio)
#             on quarterly data
x2<-cbind(retail=rjd3toolkit::aggregate(x[,1],4), clothing=rjd3toolkit::aggregate(x[,3],4))
outliers = list(retail=c("20081001"=10, "20090101"=10),clothing=c("20081001"=10))
rslt2a<-multiTD(Y, x2, model="mbDenton", outliers=outliers)
# \dontrun{
# runShiny(rslt2a)
# }

# Example 2b: Model-based Denton with outliers and fixed disaggregated BI ratio
# (adjustments based on the results of 2a)
outliers = list(retail=c("20081001"=10))
disagBIfixed = list(retail=c("20080101"=7.246695, "20080401"=7.246695,
                             "20080701"=7.246695, "20081001"=6.0))
rslt2b<-multiTD(Y, x2, model="mbDenton", outliers=outliers, disagBIfixed=disagBIfixed)
# \dontrun{
# runShiny(rslt2b)
# }

# Example 3: Enhanced model-based Denton (i.e. use of forecast of the annual
# BI ratio in extrapolation). In this example, user-defined forecast for
# the series 'retail' and automatic forecast for the series 'clothing'
Y3<-window(Y, end=2009)
x3<-window(x2, end=c(2010,2))
forecastBI.values = list(retail=c("Y+1"=6.8, "Y+2"=7.0))
rslt3<-multiTD(Y3, x3, model="mbDenton", forecastBI = "userdefined+auto",
               forecastBI.values=forecastBI.values, forecastBI.quantile="q.99")

}
