% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{runShiny}
\alias{runShiny}
\title{Run Shiny App on the results}
\usage{
runShiny(rslt)
}
\arguments{
\item{rslt}{an object of class 'nbb.multiTD.output'}
}
\description{
Visualization of the results of class 'nbb.multiTD.output'
}
