% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{summary.nbb.multiTD.output}
\alias{summary.nbb.multiTD.output}
\title{Summary function for objects of class 'nbb.multiTD.output'}
\usage{
\method{summary}{nbb.multiTD.output}(object, ...)
}
\arguments{
\item{object}{an object of class 'nbb.multiTD.output'}

\item{\dots}{further elements to display.}
}
\description{
Summary function for objects of class 'nbb.multiTD.output'
}
